Load spec files from filesystem and run with Cypress Module API
Using the Cypress Module API, tests are loaded from the filesystem and
passed to the runner. For now, the Cypress config is assumed to live in
the parent of the scripts working-directory. The user's Cypress
configuration file (cypress.json) will be used when running test.

Results and errors are captured and will be properly logged in a
follow-up commit.