Print formatted table of test results
The Cypress run results are massaged and mangled into a user-displayable
format, creating a summation of the success/failure/skipped statistics
from each unique spec file.

In the future and with minimal modifications, the stress-test tool should
be capable of accepting multiple spec names to run as part of the test.