Add proper handling for Cypress configuration and integration test files
Proper handlers wrap file system access for the Cypress configuration
file (cypress.json) and integration test (*.spec.js) files.

Following the official Cypress test runner, the `--configFile`
command-line option can be used to specify the path to a user's Cypress
configuration file. A boolean value of 'false' can be specified to
disable loading a configuration file.

When a configuration file is not specified _or_ found, the runner still
needs to know where to find the user-specified test files. The
`--integrationFolder` command-line option can be used to specify a path
to the integration test files.