Add new `parallel-run` command to run test suite in parallelized runners
A new command has been added to allow users to run their entire test
suite in a set of parallel runners. The number of concurrent runners can
be specified using the same `--threads` command-line option as the
stress test command.

To support a second command, many steps of the application required
abstraction.