Allow both commands to take one or more test file identifiers
For both the `stress-test` and `run-parallel` commands, the Cypress spec
file identifier can now contain one or more space-separated strings;
rather than the previous limitation of running a single file.

Here's an example of this being used with the stress test command:

  ```shell
    $ cypress-utils stress-test spec1 spec2
  ```

Both of these tests will be included in the stress test process, if
found in the specified integration folder.