Safely handle failures when processing results
When the Cypress Module API's `Cypress#run` call fails, the results
object returned is only a subset of the 'success' object. The only
reliable way I've found to filter these is to check for the existence of
a `failures` key, which is one of the two properties in a failed result.

While failures are filtered, the expected failures will still show up in
the accumulated stress test results. The failures filtered here are due
to Cypress-related issues, and not an issue in user's application or
test code.