Fix fall-back behavior for `--integrationFolder` command-line option
The `--integrationFolder` command-line option specifies the path where
Cypress tests are located. The specified file identifiers are searched
for within this path.

In the scenario where the user-specified configuration file succeeded to
load, but did not contain an `integrationFolder` property, having the
value fall-back behavior within the catch block was not sufficient.

This handler has been moved outside of the error boundary and runs for
all results.